requirements.txt
imageio==2.8.0
matplotlib==3.1.3
numpy==1.19.2
pillow==8.1.0
scipy==1.4.1
tensorboard==1.13.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tqdm==4.46.0
